requirements.txt
# Core MLX and ML packages
mlx>=0.15.0
mlx-lm>=0.15.0

# Data handling
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Scientific computing
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# HuggingFace integration
huggingface-hub>=0.17.0
transformers>=4.30.0
tokenizers>=0.13.0

# Optional evaluation packages (install manually if needed)
# evaluate>=0.4.0
# bert-score>=0.3.13

# Development and jupyter support
jupyter>=1.0.0
ipywidgets>=8.0.0

# File I/O
safetensors>=0.3.0